+++ src/main/java/project/MainVariables.java
@@ -9,6 +9,6 @@ public class MainVariables {
     public static long pathLength = 10;
     public static long timeIntervals = 100;
     public static int passiveIncome = 1;
-    public static int startingMoney = 100;
-    public static double gridSize = (double) (sizeY / 10);
+    public static int startingMoney=100;
+    public static double gridSize=(sizeY/10);
 }
